LPS-15051 Manual
git-svn-id: svn://svn.liferay.com/repos/public/plugins/trunk@71692 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portlets/knowledge-base-portlet/docroot/WEB-INF/src/com/liferay/knowledgebase/service/impl/ArticleServiceImpl.java
@@ -61,7 +61,7 @@ public class ArticleServiceImpl extends ArticleServiceBaseImpl {
 
 	public Article addArticle(
 			long parentResourcePrimKey, String title, String content,
-			String description, int priority, String dirName,
+			String description, long priority, String dirName,
 			ServiceContext serviceContext)
 		throws PortalException, SystemException {
 
@@ -399,7 +399,7 @@ public class ArticleServiceImpl extends ArticleServiceBaseImpl {
 
 	public Article updateArticle(
 			long resourcePrimKey, long parentResourcePrimKey, String title,
-			String content, String description, int priority, String dirName,
+			String content, String description, long priority, String dirName,
 			ServiceContext serviceContext)
 		throws PortalException, SystemException {
